AWS/requirements.txt
requests
openai
python-dotenv
redis
sentence-transformers
numpy